@@ -4,11 +4,11 @@ module SubmissionsHelper
     data_file = data_file_param.read unless data_file_param.blank?
 
     unless data_dictionary.blank?
-      dd_processor = EMERGE::Phenotype::DataDictionaryValidator.new data_dictionary, :csv
+      dd_processor = EMERGE::Phenotype::DataDictionaryValidator.new data_dictionary, :csv, ValidationParameters['error_limit']
       @data_dictionary_results = dd_processor.validate
       @data_dictionary_results[:file_name] = data_dictionary_param.original_filename
       unless data_file.blank?
-        df_processor = EMERGE::Phenotype::DataFileValidator.new data_file, dd_processor.variables, :csv
+        df_processor = EMERGE::Phenotype::DataFileValidator.new data_file, dd_processor.variables, :csv, ValidationParameters['error_limit']
         @data_file_results = df_processor.validate
         @data_file_results[:file_name] = data_file_param.original_filename
       end
