@@ -8,7 +8,6 @@ end
 
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
-require 'rspec/autorun'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
@@ -33,4 +32,6 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = true
 
   config.fixture_path = "#{::Rails.root}/test/fixtures"
+
+  config.include Rails.application.routes.url_helpers
 end
